Remove additional transcation comment
