+++ Helper/RequestHelper.php
@@ -4,7 +4,8 @@ namespace Omise\Payment\Helper;
 
 use Magento\Framework\App\RequestInterface;
 use Magento\Framework\HTTP\Header;
-use Omise\Payment\Gateway\Http\Client\APMSession;
+use Omise\Payment\Model\Omise;
+use OmiseException;
 
 class RequestHelper
 {
@@ -15,9 +16,19 @@ class RequestHelper
     private $request;
 
     /**
-     * @var APMSession
+     * @var Omise
      */
-    private $apmSession;
+    private $omise;
+
+    /**
+     * @var int
+     */
+    private $omiseConnectTimeout = 30;
+
+    /**
+     * @var int
+     */
+    private $omiseTimeout = 60;
 
     /**
      * @var \Magento\Framework\HTTP\Header
@@ -27,15 +38,15 @@ class RequestHelper
     /**
      * @param RequestInterface $request
      * @param Header $header
-     * @param APMSession $apmSession
+     * @param Omise $omise
      */
     public function __construct(
         RequestInterface $request,
         Header $header,
-        APMSession $apmSession
+        Omise $omise
     ) {
         $this->request = $request;
-        $this->apmSession = $apmSession;
+        $this->omise = $omise;
         $this->header = $header;
     }
 
@@ -70,17 +81,6 @@ class RequestHelper
 
         // return default remote IP address
         return $this->request->getServer('REMOTE_ADDR', '');
-    }
-
-    private function processForwardedForHeader($forwardedForHeader)
-    {
-        // Split if multiple IP addresses exist and get the last IP address
-        if (strpos($forwardedForHeader, ',') !== false) {
-            $multiple_ips = explode(",", $forwardedForHeader);
-            return trim(current($multiple_ips));
-        }
-
-        return $forwardedForHeader;
     }
 
     /**
@@ -119,14 +119,148 @@ class RequestHelper
      * @param bool $is_json
      * @return array
      */
-    public function sendSessionRequest($url, $requestMethod, $skey, $params = [], $is_json = false)
+    public function sendUpaSessionRequest($url, $requestMethod, $skey, $params = [], $is_json = false)
     {
-        return $this->apmSession->request(
+        return $this->upaRequest(
             $url,
             $requestMethod,
             $skey,
             $params,
             $is_json
         );
+    }
+
+    /**
+     * @param  string $url
+     * @param  string $requestMethod
+     * @param  string $key
+     * @param  array  $params
+     * @param  bool   $is_json
+     * @return array
+     */
+    private function upaRequest($url, $requestMethod, $key, $params = null, $is_json = false)
+    {
+        try {
+            $response = $this->execute($url, $requestMethod, $key, $params, $is_json);
+            $array = json_decode($response, true);
+            // If response is invalid or not a JSON.
+            if (!$this->isValidAPIResponse($array)) {
+                throw new \Exception('Unknown error. (Bad Response)');
+            }
+
+            if (!empty($array['object']) && $array['object'] === 'error') {
+                throw \OmiseException::getInstance($array);
+            }
+            return $array;
+        } catch (OmiseException $e) {
+            throw new OmiseException($e->getMessage());
+        }
+    }
+    /**
+     * @param  string $url
+     * @param  string $requestMethod
+     * @param  string $key
+     * @param  array  $params
+     * @param  bool   $is_json
+     * @return string
+     */
+    private function execute($url, $requestMethod, $key, $params = null, $is_json = false)
+    {
+        $ch = curl_init($url);
+
+        curl_setopt_array($ch, $this->genOptions($requestMethod, $key . ':', $params, $is_json));
+
+        // Make a request or thrown an exception.
+        if (($result = curl_exec($ch)) === false) {
+            $error = curl_error($ch);
+            curl_close($ch);
+
+            throw new \Exception($error);
+        }
+
+        // Close.
+        curl_close($ch);
+
+        return $result;
+    }
+
+    /**
+     * Creates an option for php-curl from the given request method and parameters in an associative array.
+     *
+     * @param  string $requestMethod
+     * @param  array  $params
+     *
+     * @return array
+     */
+    private function genOptions($requestMethod, $userpwd, $params, $is_json)
+    {
+        $options = [
+            // Set the HTTP version to 1.1.
+            CURLOPT_HTTP_VERSION => CURL_HTTP_VERSION_1_1,
+            // Set the request method.
+            CURLOPT_CUSTOMREQUEST => $requestMethod,
+            // Make php-curl returns the data as string.
+            CURLOPT_RETURNTRANSFER => true,
+            // Do not include the header in the output.
+            CURLOPT_HEADER => false,
+            // Track the header request string and set the referer on redirect.
+            CURLINFO_HEADER_OUT => true,
+            CURLOPT_AUTOREFERER => true,
+            // Make HTTP error code above 400 an error.
+            // CURLOPT_FAILONERROR => true,
+            // Time before the request is aborted.
+            CURLOPT_TIMEOUT => $this->omiseTimeout,
+            // Time before the request is aborted when attempting to connect.
+            CURLOPT_CONNECTTIMEOUT => $this->omiseConnectTimeout,
+            // Authentication.
+            CURLOPT_USERPWD => $userpwd
+        ];
+
+        // Config UserAgent
+        if ( defined('OMISE_USER_AGENT_SUFFIX') ) {
+            $options += [CURLOPT_USERAGENT => OMISE_USER_AGENT_SUFFIX];
+        } else {
+            $this->omise->defineUserAgent();
+            $options += [CURLOPT_USERAGENT => OMISE_USER_AGENT_SUFFIX];
+        }
+
+        if ($is_json) {
+            $options[CURLOPT_HTTPHEADER][] = 'Content-Type: application/json';
+            $http_query = json_encode($params);
+            $options += [CURLOPT_POSTFIELDS => $http_query];
+            return $options;
+        }
+
+        // Also merge POST parameters with the option.
+        if (is_array($params) && count($params) > 0) {
+            $http_query = http_build_query($params);
+            $http_query = preg_replace('/%5B\d+%5D/simU', '%5B%5D', $http_query);
+
+            $options += [CURLOPT_POSTFIELDS => $http_query];
+        }
+        return $options;
+    }
+
+    /**
+     * Checks if response from API was valid.
+     *
+     * @param  array  $array  - decoded JSON response
+     *
+     * @return boolean
+     */
+    private static function isValidAPIResponse($array)
+    {
+        return $array && count($array) && isset($array['object']);
+    }
+
+    private function processForwardedForHeader($forwardedForHeader)
+    {
+        // Split if multiple IP addresses exist and get the last IP address
+        if (strpos($forwardedForHeader, ',') !== false) {
+            $multiple_ips = explode(",", $forwardedForHeader);
+            return trim(current($multiple_ips));
+        }
+
+        return $forwardedForHeader;
     }
 }
